Updated vagrant setup to include authentication
- Added keyfile for testing
- Updated start command
- Made some additional changes to avoid overwriting certain files
  when they already existing on the target VM

+++ ssh-test-setup/install-leiningen.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 mkdir -p /home/vagrant/bin
-wget https://raw.githubusercontent.com/technomancy/leiningen/stable/bin/lein -O /home/vagrant/bin/lein
-chown vagrant:vagrant /home/vagrant/bin/lein
-chmod a+x /home/vagrant/bin/lein
+if [ ! -f /home/vagrant/bin/lein ]; then
+    wget https://raw.githubusercontent.com/technomancy/leiningen/stable/bin/lein -O /home/vagrant/bin/lein
+    chown vagrant:vagrant /home/vagrant/bin/lein
+    chmod a+x /home/vagrant/bin/lein
+fi
